+++ janus_server/Cargo.toml
@@ -20,7 +20,6 @@ futures = "0.3.21"
 hex = "0.4.3"
 hpke = { version = "0.8.0", features = ["default", "std"] }
 http = "0.2.6"
-lazy_static = "1"
 num_enum = "0.5.6"
 postgres-types = { version = "0.2.2", features = ["derive"] }
 prio = { git = "https://github.com/divviup/libprio-rs", rev = "229cd9c45924c7dae3ba754a6eb46b2a05ca8451" }  # TODO(brandon): use a numbered version, once a release is cut >0.7.0
@@ -33,7 +32,6 @@ serde_yaml = "0.8.23"
 signal-hook = "0.3.13"
 signal-hook-tokio = { version = "0.3.1", features = ["futures-v0_3"] }
 structopt = "0.3.26"
-testcontainers = "0.13.0"
 thiserror = "1.0"
 tokio = { version = "^1.9", features = ["full", "tracing"] }
 tokio-postgres = { version = "0.7.5", features = ["with-chrono-0_4"] }
@@ -47,8 +45,11 @@ warp = { version = "^0.3", features = ["tls"] }
 assert_matches = "1"
 hex = { version = "0.4.3", features = ["serde"] }
 hyper = "0.14.17"
+lazy_static = "1"
 libc = "0.2.123"
 mockito = "0.31.0"
 tempfile = "3.3.0"
+testcontainers = "0.13.0"
+test_util = { path = "../test_util" }
 trycmd = "0.13.3"
 wait-timeout = "0.2.0"
