+++ monolithic_integration_test/Cargo.toml
@@ -5,8 +5,14 @@ edition = "2021"
 
 [dev-dependencies]
 chrono = "0.4.19"
+deadpool-postgres = "0.10.1"
 futures = "0.3.21"
 janus_server = { path = "../janus_server" }
+lazy_static = "1"
 prio = { git = "https://github.com/divviup/libprio-rs", rev = "229cd9c45924c7dae3ba754a6eb46b2a05ca8451" }  # TODO(brandon): use a numbered version, once a release is cut >0.7.0
+ring = "0.16.20"
+testcontainers = "0.13.0"
+test_util = { path = "../test_util" }
 tokio = { version = "^1.9", features = ["full", "tracing"] }
+tokio-postgres = { version = "0.7.5", features = ["with-chrono-0_4"] }
 url = { version = "2.2.2", features = ["serde"] }
